+++ src/Models/MetadataTrait.php
@@ -12,11 +12,13 @@ use AlgoWeb\PODataLaravel\Query\LaravelReadQuery;
 use Illuminate\Database\Eloquent\Model;
 use Illuminate\Database\Eloquent\Relations\BelongsTo;
 use Illuminate\Database\Eloquent\Relations\BelongsToMany;
+use Illuminate\Database\Eloquent\Relations\HasManyThrough;
 use Illuminate\Database\Eloquent\Relations\MorphMany;
 use Illuminate\Database\Eloquent\Relations\MorphOne;
 use Illuminate\Database\Eloquent\Relations\MorphToMany;
 use Illuminate\Database\Eloquent\Relations\Relation;
 use Illuminate\Support\Facades\App;
+use Mockery\Mock;
 
 trait MetadataTrait
 {
@@ -114,10 +116,8 @@ trait MetadataTrait
         $visible = $this->getVisible();
         $hidden = $this->getHidden();
         if (0 < count($visible)) {
-            assert(!empty($visible));
             $attribs = array_intersect($visible, $attribs);
         } elseif (0 < count($hidden)) {
-            assert(!empty($hidden));
             $attribs = array_diff($attribs, $hidden);
         }
 
@@ -127,13 +127,14 @@ trait MetadataTrait
     /*
      * Get the endpoint name being exposed
      *
+     * @return null|string;
      */
     public function getEndpointName()
     {
         $endpoint = isset($this->endpoint) ? $this->endpoint : null;
 
         if (!isset($endpoint)) {
-            $bitter = get_class();
+            $bitter = get_class($this);
             $name = substr($bitter, strrpos($bitter, '\\')+1);
             return ($name);
         }
@@ -233,8 +234,8 @@ trait MetadataTrait
                             'Function definition must have keyword \'function\''
                         );
                         $begin = strpos($code, 'function(');
-                        $code = substr($code, $begin, strrpos($code, '}')-$begin+1);
-                        $lastCode = $code[strlen($code)-1];
+                        $code = substr($code, /** @scrutinizer ignore-type */$begin, strrpos($code, '}')-$begin+1);
+                        $lastCode = $code[strlen(/** @scrutinizer ignore-type */$code)-1];
                         assert('}' == $lastCode, 'Final character of function definition must be closing brace');
                         foreach ([
                                      'hasMany',
@@ -249,7 +250,7 @@ trait MetadataTrait
                                      'morphedByMany'
                                  ] as $relation) {
                             $search = '$this->' . $relation . '(';
-                            if ($pos = stripos($code, $search)) {
+                            if (stripos(/** @scrutinizer ignore-type */$code, $search)) {
                                 //Resolve the relation's model to a Relation object.
                                 $relationObj = $model->$method();
                                 if ($relationObj instanceof Relation) {
@@ -371,7 +372,7 @@ trait MetadataTrait
 
         foreach ($getterz as $getter) {
             $residual = substr($getter, 3);
-            $residual = substr($residual, 0, -9);
+            $residual = substr(/** @scrutinizer ignore-type */$residual, 0, -9);
             $methods[] = $residual;
         }
         return $methods;
@@ -428,7 +429,7 @@ trait MetadataTrait
 
     private function polyglotKeyMethodBackupNames($foo, $condition = false)
     {
-        $fkList = ['getForeignKey', 'getForeignKeyName'];
+        $fkList = ['getForeignKey', 'getForeignKeyName', 'getQualifiedFarKeyName'];
         $rkList = ['getOtherKey', 'getQualifiedParentKeyName'];
 
         $fkMethodName = null;
@@ -461,6 +462,16 @@ trait MetadataTrait
         return [$fkMethodName, $rkMethodName];
     }
 
+    private function polyglotThroughKeyMethodNames(HasManyThrough $foo)
+    {
+        $thruList = ['getThroughKey', 'getQualifiedFirstKeyName'];
+
+        $methodList = get_class_methods(get_class($foo));
+        $thruCombo = array_values(array_intersect($thruList, $methodList));
+        return $thruCombo[0];
+    }
+
+
     /**
      * @param             $hooks
      * @param             $first
@@ -471,7 +482,7 @@ trait MetadataTrait
      * @param string|null $targ
      * @param null|mixed  $type
      */
-    private function addRelationsHook(&$hooks, $first, $property, $last, $mult, $targ, $type = null)
+    private function addRelationsHook(&$hooks, $first, $property, $last, $mult, $targ, $type = null, $through = null)
     {
         if (!isset($hooks[$first])) {
             $hooks[$first] = [];
@@ -482,6 +493,7 @@ trait MetadataTrait
         $hooks[$first][$targ][$property] = [
             'property' => $property,
             'local' => $last,
+            'through' => $through,
             'multiplicity' => $mult,
             'type' => $type
         ];
@@ -497,22 +509,24 @@ trait MetadataTrait
             if ($foo instanceof MorphMany || $foo instanceof MorphToMany) {
                 continue;
             }
-            $isBelong = $foo instanceof BelongsToMany;
             $mult = '*';
             $targ = get_class($foo->getRelated());
+            list($thruName, $fkMethodName, $rkMethodName) = $this->getRelationsHasManyKeyNames($foo);
 
-            list($fkMethodName, $rkMethodName) = $this->polyglotKeyMethodNames($foo, $isBelong);
-            list($fkMethodAlternate, $rkMethodAlternate) = $this->polyglotKeyMethodBackupNames($foo, !$isBelong);
-
-            $keyRaw = $isBelong ? $foo->$fkMethodName() : $foo->$fkMethodAlternate();
+            $keyRaw = $foo->$fkMethodName();
             $keySegments = explode('.', $keyRaw);
             $keyName = $keySegments[count($keySegments)-1];
-            $localRaw = $isBelong ? $foo->$rkMethodName() : $foo->$rkMethodAlternate();
+            $localRaw = $foo->$rkMethodName();
             $localSegments = explode('.', $localRaw);
             $localName = $localSegments[count($localSegments)-1];
+            if (null !== $thruName) {
+                $thruRaw = $foo->$thruName();
+                $thruSegments = explode('.', $thruRaw);
+                $thruName = $thruSegments[count($thruSegments)-1];
+            }
             $first = $keyName;
             $last = $localName;
-            $this->addRelationsHook($hooks, $first, $property, $last, $mult, $targ);
+            $this->addRelationsHook($hooks, $first, $property, $last, $mult, $targ, null, $thruName);
         }
     }
 
@@ -619,7 +633,7 @@ trait MetadataTrait
      */
     public function getEagerLoad()
     {
-        assert(is_array($this->loadEagerRelations));
+        assert(is_array($this->loadEagerRelations), 'LoadEagerRelations not an array');
         return $this->loadEagerRelations;
     }
 
@@ -721,7 +735,10 @@ trait MetadataTrait
                     $stub->setForeignField($targType ? $key : null);
                     $stub->setMultiplicity($multArray[$relGubbins['multiplicity']]);
                     $stub->setTargType($targType);
-                    assert($stub->isOk());
+                    if (null !== $relGubbins['through']) {
+                        $stub->setThroughField($relGubbins['through']);
+                    }
+                    assert($stub->isOk(), 'Generated stub not consistent');
                     $stubs[$property] = $stub;
                 }
             }
@@ -737,6 +754,9 @@ trait MetadataTrait
             return;
         }
         $fieldName = LaravelReadQuery::PK;
+        /* As MetadataTrait is assumed to be deployed on top of an Eloquent model, which has getKey(), this is a
+           false positive */
+        /** @scrutinizer ignore-call */
         $this->$fieldName = $this->getKey();
     }
 
@@ -775,5 +795,21 @@ trait MetadataTrait
         self::$tableData = [];
         self::$tableColumnsDoctrine = [];
         self::$tableColumns = [];
+    }
+
+    private function getRelationsHasManyKeyNames($foo)
+    {
+        $thruName = null;
+        if ($foo instanceof HasManyThrough) {
+            list($fkMethodName, $rkMethodName) = $this->polyglotKeyMethodBackupNames($foo, true);
+            $thruName = $this->polyglotThroughKeyMethodNames($foo);
+            return array($thruName, $fkMethodName, $rkMethodName);
+        } elseif ($foo instanceof BelongsToMany) {
+            list($fkMethodName, $rkMethodName) = $this->polyglotKeyMethodNames($foo, true);
+            return array($thruName, $fkMethodName, $rkMethodName);
+        } else {
+            list($fkMethodName, $rkMethodName) = $this->polyglotKeyMethodBackupNames($foo, true);
+            return array($thruName, $fkMethodName, $rkMethodName);
+        }
     }
 }