+++ tests/TestMonomorphicParentOfMorphTarget.php
@@ -69,4 +69,14 @@ class TestMonomorphicParentOfMorphTarget extends Model
     {
         return $this->hasMany(TestMorphTarget::class, 'child_id');
     }
+
+    public function monomorphicChildren()
+    {
+        return $this->hasManyThrough(
+            TestMonomorphicChildOfMorphTarget::class,
+            TestMorphTarget::class,
+            'parent_id',
+            'child_id'
+        );
+    }
 }