+++ tests/unit/Models/MetadataGubbinsHolderTest.php
@@ -346,4 +346,41 @@ class MetadataGubbinsHolderTest extends TestCase
         $this->assertTrue($result[0] instanceof AssociationMonomorphic, get_class($result[0]));
         $this->assertEquals(1, count($result[0]->getLast()));
     }
+
+    public function testGetGubbinsIncludingHasManyThroughRelation()
+    {
+        $metaRaw['id'] = ['type' => 'integer', 'nullable' => false, 'fillable' => false, 'default' => null];
+        $metaRaw['alternate_id'] = ['type' => 'integer', 'nullable' => false, 'fillable' => false, 'default' => null];
+        $metaRaw['name'] = ['type' => 'string', 'nullable' => false, 'fillable' => true, 'default' => null];
+
+        $model = new TestMonomorphicParentOfMorphTarget($metaRaw);
+
+        $gubbins = $model->extractGubbins();
+        $stubs = $gubbins->getStubs();
+        $this->assertTrue(array_key_exists('monomorphicChildren', $stubs));
+    }
+
+    public function testGetBidirectionalHasManyThroughRelation()
+    {
+        $metaRaw['id'] = ['type' => 'integer', 'nullable' => false, 'fillable' => false, 'default' => null];
+        $metaRaw['alternate_id'] = ['type' => 'integer', 'nullable' => false, 'fillable' => false, 'default' => null];
+        $metaRaw['name'] = ['type' => 'string', 'nullable' => false, 'fillable' => true, 'default' => null];
+
+        $model = new TestMonomorphicParentOfMorphTarget($metaRaw);
+        $nuModel = new TestMonomorphicChildOfMorphTarget($metaRaw);
+
+        $modelGubbins = $model->extractGubbins();
+        $nuModelGubbins = $nuModel->extractGubbins();
+
+        $left = $modelGubbins->getStubs()['monomorphicChildren'];
+        $right = $nuModelGubbins->getStubs()['monomorphicParent'];
+        $this->assertTrue($left->isCompatible($right));
+
+        $foo = new MetadataGubbinsHolder();
+        $foo->addEntity($model->extractGubbins());
+        $foo->addEntity($nuModel->extractGubbins());
+
+        $result = $foo->getRelationsByRelationName(TestMonomorphicParentOfMorphTarget::class, 'monomorphicChildren');
+        $this->assertTrue(0 < count($result));
+    }
 }